+++ python/ray/data/datasource/__init__.py
@@ -1,5 +1,9 @@
 from ray.data.datasource.bigquery_datasource import BigQueryDatasource
 from ray.data.datasource.binary_datasource import BinaryDatasource
+from ray.data.datasource.block_path_provider import (
+    BlockWritePathProvider,
+    DefaultBlockWritePathProvider,
+)
 from ray.data.datasource.csv_datasource import CSVDatasource
 from ray.data.datasource.datasource import (
     Datasource,
@@ -11,8 +15,6 @@ from ray.data.datasource.datasource import (
     WriteResult,
 )
 from ray.data.datasource.file_based_datasource import (
-    BlockWritePathProvider,
-    DefaultBlockWritePathProvider,
     FileBasedDatasource,
     FileExtensionFilter,
     _S3FileSystemWrapper,
